refactor(gda): extract CLI dispatch tails into gda.dispatch (ADR-0040)
Pure move of _emit/_dispatch*/_run_params_json and the runner seams out
of cli.py; recipes reference the seams via late-bound module attributes
so test monkeypatches keep binding. Test patch targets follow the move.
