+++ tests/test_export_run_commands.py
@@ -65,12 +65,14 @@ def _inject(monkeypatch, *, get=GET_RESULT, export=None):
             exit_code=0,
         )
     )
-    monkeypatch.setattr("gda.cli._make_runner", lambda binary, project=None: get_runner)
+    monkeypatch.setattr(
+        "gda.dispatch._make_runner", lambda binary, project=None: get_runner
+    )
     if export is None:
         export = RunResult(stdout="", stderr="", exit_code=0)
     export_runner = FakeExportRunner(export)
     monkeypatch.setattr(
-        "gda.cli._make_export_runner", lambda binary, project=None: export_runner
+        "gda.dispatch._make_export_runner", lambda binary, project=None: export_runner
     )
     return get_runner, export_runner
 
@@ -254,8 +256,8 @@ def test_export_run_rejects_unknown_mode(monkeypatch, tmp_path):
     def _boom(*args, **kwargs):
         raise AssertionError("a rejected --mode must not spawn any engine")
 
-    monkeypatch.setattr("gda.cli._make_runner", _boom)
-    monkeypatch.setattr("gda.cli._make_export_runner", _boom)
+    monkeypatch.setattr("gda.dispatch._make_runner", _boom)
+    monkeypatch.setattr("gda.dispatch._make_export_runner", _boom)
 
     result = CliRunner().invoke(
         app,
@@ -634,10 +636,12 @@ def test_export_run_unknown_preset_reuses_export_get_error(monkeypatch, tmp_path
             exit_code=4,
         )
     )
-    monkeypatch.setattr("gda.cli._make_runner", lambda binary, project=None: get_runner)
+    monkeypatch.setattr(
+        "gda.dispatch._make_runner", lambda binary, project=None: get_runner
+    )
     export_runner = FakeExportRunner(RunResult(stdout="", stderr="", exit_code=0))
     monkeypatch.setattr(
-        "gda.cli._make_export_runner", lambda binary, project=None: export_runner
+        "gda.dispatch._make_export_runner", lambda binary, project=None: export_runner
     )
 
     result = CliRunner().invoke(
@@ -658,8 +662,8 @@ def test_export_run_schema_emits_contract_without_engine(monkeypatch):
     def _boom(*args, **kwargs):
         raise AssertionError("--schema must not spawn any engine")
 
-    monkeypatch.setattr("gda.cli._make_runner", _boom)
-    monkeypatch.setattr("gda.cli._make_export_runner", _boom)
+    monkeypatch.setattr("gda.dispatch._make_runner", _boom)
+    monkeypatch.setattr("gda.dispatch._make_export_runner", _boom)
 
     result = CliRunner().invoke(app, ["export", "run", "--schema"])
