+++ tests/test_scene_commands.py
@@ -117,7 +117,7 @@ def test_scene_get_passes_resolved_project_to_the_runner(monkeypatch, tmp_path):
             RunResult(stdout=sentinel(GET_RESULT), stderr="", exit_code=0)
         )
 
-    monkeypatch.setattr("gda.cli._make_runner", record)
+    monkeypatch.setattr("gda.dispatch._make_runner", record)
 
     result = CliRunner().invoke(
         app, ["scene", "get", "res://main.tscn", "--project", str(tmp_path), "--json"]
@@ -289,7 +289,7 @@ def test_scene_list_passes_resolved_project_to_the_runner(monkeypatch, tmp_path)
             RunResult(stdout=sentinel(LIST_RESULT), stderr="", exit_code=0)
         )
 
-    monkeypatch.setattr("gda.cli._make_runner", record)
+    monkeypatch.setattr("gda.dispatch._make_runner", record)
 
     result = CliRunner().invoke(
         app, ["scene", "list", "--project", str(tmp_path), "--json"]
