+++ tests/test_script_commands.py
@@ -197,7 +197,7 @@ def test_script_list_passes_resolved_project_to_the_runner(monkeypatch, tmp_path
             RunResult(stdout=sentinel(LIST_RESULT), stderr="", exit_code=0)
         )
 
-    monkeypatch.setattr("gda.cli._make_runner", record)
+    monkeypatch.setattr("gda.dispatch._make_runner", record)
 
     result = CliRunner().invoke(
         app, ["script", "list", "--project", str(tmp_path), "--json"]
