+++ tests/test_skill_command.py
@@ -219,7 +219,7 @@ def test_skill_spawns_no_godot(monkeypatch):
         raise AssertionError("gda skill must not touch the engine")
 
     monkeypatch.setattr("gda.headless.resolve_godot_binary", boom)
-    monkeypatch.setattr("gda.cli._make_runner", boom)
+    monkeypatch.setattr("gda.dispatch._make_runner", boom)
 
     result = CliRunner().invoke(app, ["skill", "--json"])
